:truck: move authentication to AuthViewModel.kt
