+++ app/src/main/java/com/rivaldy/id/dicoding/ui/auth/register/RegisterActivity.kt
@@ -14,13 +14,15 @@ import com.rivaldy.id.core.utils.UtilExtensions.showSnackBar
 import com.rivaldy.id.dicoding.R
 import com.rivaldy.id.dicoding.databinding.ActivityRegisterBinding
 import com.rivaldy.id.dicoding.ui.MainViewModel
+import com.rivaldy.id.dicoding.ui.auth.AuthViewModel
 import com.rivaldy.id.dicoding.ui.home.HomeActivity
 import dagger.hilt.android.AndroidEntryPoint
 
 
 @AndroidEntryPoint
 class RegisterActivity : BaseActivity<ActivityRegisterBinding>() {
-    private val viewModel by viewModels<MainViewModel>()
+    private val viewModel by viewModels<AuthViewModel>()
+    private val mainViewModel by viewModels<MainViewModel>()
 
     override fun getViewBinding() = ActivityRegisterBinding.inflate(layoutInflater)
 
@@ -77,7 +79,7 @@ class RegisterActivity : BaseActivity<ActivityRegisterBinding>() {
 
     private fun showLoginSuccess(loginResult: LoginResult?) {
         showLoading(false)
-        viewModel.setLoginInfo(LoginInfo(loginResult?.userId, loginResult?.token, binding.emailET.text.toString(), loginResult?.name))
+        mainViewModel.setLoginInfo(LoginInfo(loginResult?.userId, loginResult?.token, binding.emailET.text.toString(), loginResult?.name))
 
         val intent = Intent(this, HomeActivity::class.java)
         intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK)
